scripts/image_to_text/requirements.txt
Pillow
pytesseract
pyperclip
pyinstaller
